% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_insulin_purchases_cols}
\alias{add_insulin_purchases_cols}
\title{Add columns for information about insulin drug purchases}
\usage{
add_insulin_purchases_cols(gld_hba1c_after_exclusions)
}
\arguments{
\item{gld_hba1c_after_exclusions}{The GLD and HbA1c data after exclusions}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}}.
Three new columns are added:
\itemize{
\item \code{has_two_thirds_insulin}: A logical variable used in classifying type 1
diabetes. See \code{\link[=algorithm]{algorithm()}} for more details.
\item \code{has_only_insulin_purchases}: A logical variable used in classifying
type 1 diabetes. See \code{\link[=algorithm]{algorithm()}} for more details.
\item \code{has_insulin_purchases_within_180_days}: A logical variable used in
classifying type 1 diabetes. See \code{\link[=algorithm]{algorithm()}} for more details.
}
}
\description{
Add columns for information about insulin drug purchases
}
\examples{
\dontrun{
simulate_registers("lmdb", 10000)[[1]] |>
  include_gld_purchases() |>
  add_insulin_purchases_cols()
}
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
