% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_majority_t1d_diagnosis_col}
\alias{add_majority_t1d_diagnosis_col}
\title{Add a column for whether the majority of diabetes diagnoses are for type 1 diabetes}
\usage{
add_majority_t1d_diagnosis_col(data)
}
\arguments{
\item{data}{Data from \code{\link[=include_diabetes_diagnoses]{include_diabetes_diagnoses()}} function.}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}},
with the following columns and up to two rows per individual:
\itemize{
\item \code{pnr}: The personal identification variable.
\item \code{dates}: The dates of the first and second hospital diabetes diagnosis.
\item \code{has_majority_t1d_diagnosis}: A logical vector indicating whether the
majority of primary diagnoses are related to type 1 diabetes.
}
}
\description{
This function evaluates whether an individual has a majority of type 1
diabetes-specific hospital diagnoses (DE10) among all type-specific diabetes
primary diagnoses (DE10 & DE11) from medical departments. If an individual
has any type-specific diabetes diagnoses from endocrinology departments,
the majority is determined only among these contacts.
}
\details{
This output is passed to the \code{join_inclusions()} function, where the
\code{dates} variable is used for the final step of the inclusion process.
The variables for whether the majority of diagnoses are for type 1 diabetes
is used for later classification of type 1 diabetes.
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
