% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-thirds-insulin.R
\name{add_two_thirds_and_only_insulin}
\alias{add_two_thirds_and_only_insulin}
\title{Add columns for when two-thirds of doses are insulin and if only insulin was purchased.}
\usage{
add_two_thirds_and_only_insulin(data, gld_purchases)
}
\arguments{
\item{data}{A \code{\link[tibble:tibble]{tibble::tibble()}} with at least the column \code{pnr}.}

\item{gld_purchases}{The output from \code{\link[=include_gld_purchases]{include_gld_purchases()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with two extra columns called
\code{is_two_thirds_insulin} and \code{only_insulin_purchases}. These columns
are logical variables indicating whether at least two-thirds of the
GLD doses purchased are insulin doses and if only insulin was
purchased.
}
\description{
Add columns for when two-thirds of doses are insulin and if only insulin was purchased.
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
