% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-thirds-insulin.R
\name{add_two_thirds_insulin_doses}
\alias{add_two_thirds_insulin_doses}
\title{Add column indicating if at least two-thirds of GLD doses are insulin doses.}
\usage{
add_two_thirds_insulin_doses(data, gld_purchases)
}
\arguments{
\item{data}{A \code{\link[tibble:tibble]{tibble::tibble()}} with at least the columns \code{pnr} and \code{date}.}

\item{gld_purchases}{The output from \code{\link[=include_gld_purchases]{include_gld_purchases()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one extra column called
\code{two_thirds_insulin_doses}. This column is a logical variable indicating
whether at least two-thirds of the GLD doses purchased are insulin doses.
}
\description{
Add column indicating if at least two-thirds of GLD doses are insulin doses.
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
