% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{algorithm}
\alias{algorithm}
\title{A list of the algorithmic logic for each criteria underlying osdc.}
\format{
Is a list with nested lists that have these named elements:

\describe{
\item{register}{Optional. The register used for this criteria.}
\item{name}{The inclusion or exclusion criteria name.}
\item{title}{The title to use when displaying the algorithmic logic in tables.}
\item{logic}{The logic for the criteria.}
\item{comments}{Some additional comments on the criteria.}
}
}
\usage{
algorithm()
}
\value{
A nested list with the algorithmic logic for each criteria. Contains
fields \code{register}, \code{title}, \code{logic}, and \code{comments}.
}
\description{
This nested list contains the logic details of the algorithm for specific
inclusion and exclusion criteria.
}
\examples{
algorithm()$hba1c
algorithm()$gld$logic
}
\seealso{
See the \code{vignette("algorithm")} and \link{algorithm} for the logic used
to filter these patients.
}
