% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{as_date}
\alias{as_date}
\title{Simple \code{as.Date()} wrapper.}
\usage{
as_date(x)
}
\arguments{
\item{x}{A character (or date) column.}
}
\value{
A Date column.
}
\description{
DuckDB doesn't support using \code{\link[lubridate:as_date]{lubridate::as_date()}}, so this is
a simple wrapper around \code{\link[=as.Date]{as.Date()}} with the correct formats.
}
\keyword{internal}
