% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{as_sql_datetime}
\alias{as_sql_datetime}
\title{Translate to SQL for datetime conversion to eventually date}
\usage{
as_sql_datetime(x)
}
\arguments{
\item{x}{A character (or date) column, in quotes.}
}
\value{
A Datetime column.
}
\description{
DuckDB doesn't support using \code{\link[lubridate:as_date]{lubridate::as_date()}}, so this
uses \code{\link[dbplyr:sql]{dbplyr::sql()}} to directly use DuckDB's \code{strptime} to
convert strings to datetimes. Afterwards, it can be converted
to dates.
}
\keyword{internal}
