% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-cases.R
\name{cases}
\alias{cases}
\title{Create a synthetic dataset of edge case inputs}
\usage{
cases()
}
\value{
A named list of 9 \code{\link[tibble:tibble]{tibble::tibble()}} objects, each representing a
different health register: \code{bef}, \code{lmdb}, \code{lpr_adm}, \code{lpr_diag},
\code{kontakter}, \code{diagnoser}, \code{sysi}, \code{sssy}, and \code{lab_forsker}.
}
\description{
This function generates a list of tibbles representing the Danish health
registers and the data necessary to run the algorithm. The dataset contains
23 individual cases (\code{pnr}s), each designed to test a specific logical branch
of the diabetes classification algorithm, including inclusion, exclusion,
censoring, and type classification rules.
}
\details{
The generated data is used in \code{testthat} tests to ensure the algorithm
behaves as expected under a wide range of conditions, but it is also intended
to be explored by users to better understand how the algorithm logic works.
}
\examples{
\dontrun{
cases()
}
}
\seealso{
See the \link{edge-case-description}  vignette for a detailed description of these cases.
}
\keyword{internal}
