% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data-types.R
\name{check_data_types}
\alias{check_data_types}
\title{Check data types}
\usage{
check_data_types(data, register, call = rlang::caller_env())
}
\arguments{
\item{data}{A data frame to check.}

\item{register}{The name of the register to check against.}

\item{call}{The environment where the function is called, so that the error
traceback gives a more meaningful location.}
}
\value{
Either TRUE if the verification passes, or an error.
}
\description{
Check data types
}
\examples{
\dontrun{
check_data_types(simulate_register("kontakter"), "kontakter")
}
}
\keyword{internal}
