% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-diabetes.R
\name{check_duckdb}
\alias{check_duckdb}
\title{Check that data is a DuckDB connection}
\usage{
check_duckdb(data, call = rlang::caller_env())
}
\arguments{
\item{data}{Data to be checked. A \code{tbl_duckdb_connection} or
\code{duckdb_connection} object.}

\item{call}{The environment of the calling function, used to make error
messages point to the user-facing function rather than this internal check.}
}
\value{
The data, if it is a DuckDB connection. Errors if not.
}
\description{
Check that data is a DuckDB connection
}
\keyword{internal}
