% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-diabetes.R
\name{classify_diabetes}
\alias{classify_diabetes}
\title{Classify diabetes status using Danish registers.}
\usage{
classify_diabetes(
  lpr,
  hsr,
  lab_forsker,
  bef,
  lmdb,
  stable_inclusion_start_date = "1998-01-01"
)
}
\arguments{
\item{lpr}{The unified LPR register, see \code{\link[=join_registers]{join_registers()}}}

\item{hsr}{The unified health services registers (SYSI and SSSY), see
\code{\link[=join_registers]{join_registers()}}}

\item{lab_forsker}{The register for laboratory results for research}

\item{bef}{The BEF table from the civil register}

\item{lmdb}{The LMDB table from the prescription register}

\item{stable_inclusion_start_date}{Cutoff date after which inclusion events
are considered true incident diabetes cases. Defaults to "1998-01-01",
which is one year after the data on pregnancy events from the Patient Register
are considered valid for dropping gestational diabetes-related purchases of
glucose-lowering drugs. This default assumes that the user is using LPR and
LMDB data from at least Jan 1 1997 onward. If the user only has access to LPR
and LMDB data from a later date, this parameter should be set to one year
after the beginning of the user's data coverage.}
}
\value{
The same object type as the input data, which would be a
\code{\link[duckplyr:duckdb_tibble]{duckplyr::duckdb_tibble()}} type object.
}
\description{
This function requires that each source of register data is represented
as a single DuckDB object in R (e.g. a connection to Parquet files).
Each DuckDB object must contain a single table covering all years of
that data source, or at least the years you have and are interested
in.
}
\seealso{
See the \code{vignette("osdc", package = "osdc")} vignette for a more
details and on how to use this function.
}
