% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-diabetes.R
\name{classify_t2d}
\alias{classify_t2d}
\title{Classify those with type 2 diabetes.}
\usage{
classify_t2d(data)
}
\arguments{
\item{data}{Joined data output from the filtering steps.}
}
\value{
The same object type as the input data, which would be a
\code{\link[duckplyr:duckdb_tibble]{duckplyr::duckdb_tibble()}} type object.
}
\description{
Classify those with type 2 diabetes.
}
\keyword{internal}
