% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{create_fake_atc}
\alias{create_fake_atc}
\title{Create a vector with random ATC codes}
\usage{
create_fake_atc(n)
}
\arguments{
\item{n}{The number of random ATC codes to generate.}
}
\value{
A character vector of ATC codes.
}
\description{
Anatomical Therapeutic Chemical (ATC) codes are unique medicine codes
based on on what organ or system it works on and how it works.
}
\examples{
\dontrun{
create_fake_atc(10)
}
}
\keyword{internal}
