% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{create_fake_date}
\alias{create_fake_date}
\title{Create fake dates}
\usage{
create_fake_date(n, from = "1977-01-01", to = lubridate::today())
}
\arguments{
\item{n}{The number of dates to generate.}

\item{from}{A date determining the first date in the interval to sample from.}

\item{to}{A date determining the last date in the interval to sample from.}
}
\value{
A vector of dates.
}
\description{
Create fake dates
}
\examples{
\dontrun{
create_fake_date(20)
create_fake_date(20, "1995-04-19", "2024-04-19")
}
}
\keyword{internal}
