% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{create_fake_hovedspeciale_ans}
\alias{create_fake_hovedspeciale_ans}
\title{Create a vector of random department specialties}
\usage{
create_fake_hovedspeciale_ans(n)
}
\arguments{
\item{n}{The number of department specialties to create.}
}
\value{
A character vector.
}
\description{
Create a vector of random department specialties
}
\examples{
\dontrun{
create_fake_hovedspeciale_ans(1000)
}
}
\keyword{internal}
