% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{create_fake_icd}
\alias{create_fake_icd}
\title{Create a vector with random ICD-8 or -10 diagnoses}
\usage{
create_fake_icd(n, date = NULL)
}
\arguments{
\item{n}{The number of ICD-8 or -10 diagnoses to generate.}

\item{date}{A date determining whether the diagnoses should be ICD-8 or ICD-10. If null, a random date will be sampled to determine which ICD revision the diagnosis should be from. In the Danish registers, ICD-10 is used after 1994.}
}
\value{
A character vector of ICD-10 diagnoses.
}
\description{
Create a vector with random ICD-8 or -10 diagnoses
}
\keyword{internal}
