% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{create_fake_icd10}
\alias{create_fake_icd10}
\title{Create a vector of random ICD-10 diagnoses}
\source{
The stored CSV is downloaded from the Danish Health Data Authority's
website at \code{medinfo.dk}
}
\usage{
create_fake_icd10(n)
}
\arguments{
\item{n}{An integer determining how many diagnoses to create.}
}
\value{
A character vector of ICD-10 diagnoses.
}
\description{
ICD-10 is the 10th revision of the International Classification of Diseases.
}
\keyword{internal}
