% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{create_fake_icd8}
\alias{create_fake_icd8}
\title{Create a vector of random ICD-8 diagnoses}
\usage{
create_fake_icd8(n)
}
\arguments{
\item{n}{The number of ICD-8 diagnoses to generate.}
}
\value{
A character vector of ICD-8 diagnoses.
}
\description{
ICD-8 is the 8th revision of the International Classification of Diseases.
}
\examples{
\dontrun{
create_fake_icd8(1)
}
}
\keyword{internal}
