% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{create_fake_npu}
\alias{create_fake_npu}
\title{Create a vector of random NPU codes}
\usage{
create_fake_npu(n)
}
\arguments{
\item{n}{The number of NPUs to create.}
}
\value{
A character vector.
}
\description{
Nomenclature for Properties and Units (NPUs) are codes that identifies
laboratory results.
}
\examples{
\dontrun{
create_fake_npu(4)
}
}
\keyword{internal}
