% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inclusion_dates.R
\name{create_inclusion_dates}
\alias{create_inclusion_dates}
\title{Create inclusion dates from all the inclusion events.}
\usage{
create_inclusion_dates(inclusions, stable_inclusion_start_date = "1998-01-01")
}
\arguments{
\item{inclusions}{Output from \code{\link[=join_inclusions]{join_inclusions()}}.}

\item{stable_inclusion_start_date}{Cutoff date after which inclusion events
are considered reliable (e.g., after changes in drug labeling or data
entry practices). Defaults to "1998-01-01" which is one year after
obstetric codes are reliable in the GLD data (since we use LPR data to
drop rows related to gestational diabetes). This limits the included
cohort to individuals with inclusion dates after this cutoff date.}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}},
along with the \code{pnr} and \code{date} columns along with the \code{atc} column from
\code{exclude_pregnancy()}, and the \code{n_t1d_endocrinology},
\code{n_t2d_endocrinology}, \code{n_t1d_medical}, and \code{n_t2d_medical} columns from
\code{include_diabetes_diagnoses()}. It also creates two new columns:
\itemize{
\item \code{raw_inclusion_date}: Date of raw inclusion, the second earliest recorded
event for each individual.
\item \code{stable_inclusion_date}: Same as raw inclusion date, but set to \code{NA} if
the raw inclusion date is before the stable inclusion start date.
}
}
\description{
This function takes the output from
\code{join_inclusions()} and defines the final inclusion dates, raw and stable
based on all inclusion event types. Since inclusion requires at least two
events (of any type), this function keeps only those with 2 or more events.
E.g., an individual with two elevated HbA1c tests followed by a
glucose-lowering drug purchase is included with the latest elevated HbA1c
test and the purchase of glucose-lowering drugs.
}
\details{
The function creates two new columns: one indicating the "raw"
inclusion date and one indicating
a "stable" inclusion date, see @returns for more details.
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
