% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{create_simulated_data}
\alias{create_simulated_data}
\title{Simulate data based on simulation definitions}
\usage{
create_simulated_data(data, n)
}
\arguments{
\item{data}{A tibble with simulation definitions.}

\item{n}{Number of observations to simulate.}
}
\value{
A tibble with simulated data.
}
\description{
Simulate data based on simulation definitions
}
\keyword{internal}
