% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_pcos}
\alias{drop_pcos}
\title{Drop rows with metformin purchases that are potentially for the treatment of polycystic ovary syndrome}
\usage{
drop_pcos(gld_purchases, bef)
}
\arguments{
\item{gld_purchases}{The output from \code{\link[=keep_gld_purchases]{keep_gld_purchases()}}.}

\item{bef}{The \code{bef} register.}
}
\value{
The same type as the input data, as a \code{\link[duckplyr:duckdb_tibble]{duckplyr::duckdb_tibble()}}.
It also has the same columns as \code{\link[=keep_gld_purchases]{keep_gld_purchases()}}, except for a logical
helper variable \code{no_pcos} that is used in later functions.
}
\description{
Takes the output from \code{\link[=keep_gld_purchases]{keep_gld_purchases()}} and \code{bef} (information on sex and date
of birth) to drop rows with metformin purchases that are potentially for the treatment
of polycystic ovary syndrome.
This function only performs a filtering operation so it outputs the same structure and
variables as the input from \code{\link[=keep_gld_purchases]{keep_gld_purchases()}}, except the addition of a logical
helper variable \code{no_pcos} that is used in later functions.
After these rows have been dropped, the output is used by \code{drop_pregnancies()}.
}
\examples{
\dontrun{
register_data <- simulate_registers(c("lmdb", "bef"), 100)
drop_pcos(
  gld_purchases = keep_gld_purchases(register_data$lmdb),
  bef = register_data$bef
)
}
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
