% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{fct_to_chr}
\alias{fct_to_chr}
\title{Convert all factor variables to character variables.}
\usage{
fct_to_chr(data)
}
\arguments{
\item{data}{A tibble or data frame.}
}
\value{
A  \code{\link[duckplyr:duckdb_tibble]{duckplyr::duckdb_tibble()}}.
}
\description{
Convert all factor variables to character variables.
}
\keyword{internal}
