% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pregnancy-dates.R
\name{get_pregnancy_dates}
\alias{get_pregnancy_dates}
\title{Simple function to get only the pregnancy event dates.}
\usage{
get_pregnancy_dates(lpr2, lpr3)
}
\arguments{
\item{lpr2}{Output from \code{join_lpr2()}.}

\item{lpr3}{Output from \code{join_lpr3()}.}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Simple function to get only the pregnancy event dates.
}
\examples{
\dontrun{
register_data <- simulate_registers(
  c("lpr_adm", "lpr_diag", "kontakter", "diagnoser"),
  n = 1000
)
lpr2 <- join_lpr2(register_data$lpr_adm, register_data$lpr_diag)
lpr3 <- join_lpr3(register_data$kontakter, register_data$diagnoser)
get_pregnancy_dates(lpr2, lpr3)
}
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
