% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-variables.R
\name{get_register_abbrev}
\alias{get_register_abbrev}
\title{Get a list of the registers' abbreviations.}
\usage{
get_register_abbrev()
}
\value{
A character string.
}
\description{
Get a list of the registers' abbreviations.
}
\keyword{internal}
