% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-variables.R
\name{get_required_variables}
\alias{get_required_variables}
\title{Get a list of required variables from a specific register.}
\usage{
get_required_variables(register)
}
\arguments{
\item{register}{The abbreviation of the register name. See list of
abbreviations in \code{\link[=get_register_abbrev]{get_register_abbrev()}}.}
}
\value{
A character vector of variable names.
}
\description{
Get a list of required variables from a specific register.
}
\examples{
get_required_variables("bef")
}
\keyword{internal}
