% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-gld-purchases.R
\name{include_gld_purchases}
\alias{include_gld_purchases}
\title{Include only glucose lowering drugs (GLD).}
\usage{
include_gld_purchases(lmdb)
}
\arguments{
\item{lmdb}{The \code{lmdb} register.}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}}.
Only glucose lowering drugs are kept, plus some columns are renamed.
}
\description{
But don't include glucose-lowering drugs that may be used for other
conditions than diabetes like GLP-RAs or dapagliflozin/empagliflozin drugs.
Since the diagnosis code data on pregnancies (see below) is insufficient to
perform censoring prior to 1997, \code{\link[=include_gld_purchases]{include_gld_purchases()}} only extracts
dates from 1997 onward by default (if Medical Birth Register data is
available to use for censoring, the extraction window can be extended).
}
\examples{
\dontrun{
simulate_registers("lmdb", 1000)[[1]] |> include_gld_purchases()
}
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
