% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-podiatrist-services.R
\name{include_podiatrist_services}
\alias{include_podiatrist_services}
\title{Include only those who have had podiatrist services}
\usage{
include_podiatrist_services(sysi, sssy)
}
\arguments{
\item{sysi}{The SYSI register.}

\item{sssy}{The SSSY register.}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}}
with two columns:
\itemize{
\item \code{pnr}: Personal identification variable.
\item \code{dates}: The dates of podiatrist services.
}
}
\description{
See \link{algorithm} for the logic used to filter these patients.
In addition to the algorithm logic, duplicates are removed and only
the two earliest dates per individual are kept.
}
\details{
The output is passed to \code{join_inclusions()} function to be joined with
the rest of the inclusion data.
}
\examples{
\dontrun{
include_podiatrist_services(register_data$sysi, register_data$sssy)
}
}
\keyword{internal}
