% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{insert_analysiscode}
\alias{insert_analysiscode}
\title{Insert additional analysis codes for HbA1c}
\usage{
insert_analysiscode(data, proportion = 0.3)
}
\arguments{
\item{data}{A tibble.}

\item{proportion}{Proportion to re-sample. Defaults to 0.3.}
}
\value{
A tibble. If a column is named \code{analysiscode}, a proportion of the
values are replaced by codes for HbA1c.
}
\description{
Insert additional analysis codes for HbA1c
}
\keyword{internal}
