% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{insert_false_metformin}
\alias{insert_false_metformin}
\title{Insert cases where metformin is used for other purposes than diabetes}
\usage{
insert_false_metformin(data, proportion = 0.05)
}
\arguments{
\item{data}{A tibble.}

\item{proportion}{Proportion to re-sample. Defaults to 0.05.}
}
\value{
A tibble. If all column names in the tibble is either \code{atc}, a
proportion of observations is re-sampled as metformin.
}
\description{
This function uses the variable \code{indo} which is the code for the underlying
condition treated by the prescribed medication.
}
\keyword{internal}
