% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{insert_specific_atc}
\alias{insert_specific_atc}
\title{Insert specific ATC codes based on a proportion}
\usage{
insert_specific_atc(data, proportion = 0.3)
}
\arguments{
\item{data}{A tibble.}

\item{proportion}{Proportion to be resampled. Defaults to 0.3.}
}
\value{
A tibble with a proportion of resampled ATC codes for columns named
'atc'
}
\description{
Insert specific ATC codes based on a proportion
}
\keyword{internal}
