% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-inclusions.R
\name{join_inclusions}
\alias{join_inclusions}
\title{Join included events.}
\usage{
join_inclusions(
  included_diabetes_diagnoses,
  included_podiatrist_services,
  hba1c_censored_pregnancy,
  gld_censored_pcos_pregnancy
)
}
\arguments{
\item{included_diabetes_diagnoses}{Output from \code{\link[=include_diabetes_diagnoses]{include_diabetes_diagnoses()}}.}

\item{included_podiatrist_services}{Output from \code{\link[=include_podiatrist_services]{include_podiatrist_services()}}.}

\item{hba1c_censored_pregnancy}{Output from \code{\link[=exclude_pregnancy]{exclude_pregnancy()}} when given
\code{hba1c} data.}

\item{gld_censored_pcos_pregnancy}{Output from \code{\link[=exclude_pregnancy]{exclude_pregnancy()}} when
given \code{gld_censored_pcos} data.}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}},
with the joined columns from the output of \code{\link[=include_diabetes_diagnoses]{include_diabetes_diagnoses()}},
\code{\link[=include_podiatrist_services]{include_podiatrist_services()}} and \code{\link[=exclude_pregnancy]{exclude_pregnancy()}}. There will be
1-8 rows per \code{pnr}.
}
\description{
This function joins the outputs from all the inclusion and exclusion
functions, by \code{pnr} and \code{dates}. Input datasets:
\itemize{
\item \code{included_diabetes_diagnoses}: Dates are the first and second hospital
diabetes diagnosis.
\item \code{included_podiatrist_services}: Dates are the first and second
diabetes-specific podiatrist record.
\item \code{hba1c_censored_pregnancy}: Dates are the first and second elevated HbA1c
test results (after censoring potential gestational diabetes).
\item \code{gld_censored_pcos_pregnancy}: Dates are the first and second purchase of a
glucose-lowering drug (after censoring potential polycystic ovary syndrome
and gestational diabetes).
}
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
