% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-lpr.R
\name{join_lpr}
\alias{join_lpr}
\title{Join the prepared LPR registers.}
\usage{
join_lpr(lpr2, lpr3f)
}
\arguments{
\item{lpr2}{The prepared LPR2 register. See \code{\link[=prepare_lpr2]{prepare_lpr2()}}.}

\item{lpr3f}{The prepared LPR3F register. See \code{\link[=prepare_lpr3f]{prepare_lpr3f()}}.}
}
\value{
The same type as the input data, as a \code{\link[duckplyr:duckdb_tibble]{duckplyr::duckdb_tibble()}},
with the following columns:
\itemize{
\item \code{pnr}: The personal identification variable.
\item \code{date}: The date of all the recorded diagnosis (renamed from
\code{d_inddto} or \code{dato_start}).
\item \code{is_primary_diagnosis}: Whether the diagnosis was a primary diagnosis.
\item \code{is_diabetes_code}: Whether the diagnosis was any type of diabetes.
\item \code{is_t1d_code}: Whether the diagnosis was T1D-specific.
\item \code{is_t2d_code}: Whether the diagnosis was T2D-specific.
\item \code{is_pregnancy_code}: Whether the person has an event related to
pregnancy like giving birth or having a miscarriage at the given date.
\item \code{is_endocrinology_dept}: Whether the diagnosis was made by an
endocrinology medical department.
\item \code{is_medical_dept}: Whether the diagnosis was made by a
non-endocrinology medical department.
}
}
\description{
Join the prepared LPR registers.
}
\examples{
register_data <- simulate_registers(c(
  "lpr_adm",
  "lpr_diag",
  "lpr3f_kontakter",
  "lpr3f_diagnoser"
))

join_lpr(
  lpr2 = prepare_lpr2(register_data$lpr_adm, register_data$lpr_diag),
  lpr3f = prepare_lpr3f(
    register_data$lpr3f_kontakter,
    register_data$lpr3f_diagnoser
  )
)
}
