% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{join_lpr2}
\alias{join_lpr2}
\title{Join together the LPR2 (\code{lpr_diag} and \code{lpr_adm}) registers.}
\usage{
join_lpr2(lpr_adm, lpr_diag)
}
\arguments{
\item{lpr_adm}{The admission register.}

\item{lpr_diag}{The diagnosis register.}
}
\value{
The same class as the input, defaults to a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Join together the LPR2 (\code{lpr_diag} and \code{lpr_adm}) registers.
}
\examples{
\dontrun{
sim_data <- simulate_registers(c("lpr_adm", "lpr_diag"), 100)
join_lpr2(sim_data$lpr_adm, sim_data$lpr_diag)
}
}
\keyword{internal}
