% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{join_lpr3}
\alias{join_lpr3}
\title{Join together the LPR3 (\code{diagnoser} and \code{kontakter}) registers.}
\usage{
join_lpr3(kontakter, diagnoser)
}
\arguments{
\item{kontakter}{The contacts register.}

\item{diagnoser}{The diagnosis register.}
}
\value{
The same class as the input, defaults to a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
To prepare the LPR3 data for the inclusion process, this function also
renames the \code{cpr} variable to \code{pnr} to match the other registers.
}
\examples{
\dontrun{
kontakter <- simulate_registers("kontakter", 100)[[1]]
diagnoser <- simulate_registers("diagnoser", 100)[[1]]
join_lpr3(
  kontakter = kontakter,
  diagnoser = diagnoser
)
}
}
\keyword{internal}
