% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-lpr.R
\name{join_registers}
\alias{join_registers}
\title{Join prepared registers.}
\usage{
join_registers(register_list)
}
\arguments{
\item{register_list}{A list of the prepared registers, from e.g.
\code{\link[=prepare_lpr2]{prepare_lpr2()}}.}
}
\value{
A single object with all rows from each register in \code{register_list}.
}
\description{
Join prepared registers.
}
\examples{
register_data <- simulate_registers(c(
  "lpr_adm",
  "lpr_diag",
  "lpr3f_kontakter",
  "lpr3f_diagnoser",
  "sssy",
  "sysi"
))

join_registers(list(
  prepare_lpr2(register_data$lpr_adm, register_data$lpr_diag),
  prepare_lpr3f(
    register_data$lpr3f_kontakter,
    register_data$lpr3f_diagnoser
  )
))

join_registers(list(register_data$sysi, register_data$sssy))
}
