% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-diabetes.R
\name{keep_core_rows_with_no_na}
\alias{keep_core_rows_with_no_na}
\title{Keep rows that have no missing values for the core variables.}
\usage{
keep_core_rows_with_no_na(data)
}
\arguments{
\item{data}{The joined inclusion data.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with rows with missing core variables removed.
}
\description{
Keep rows that have no missing values for the core variables.
}
\keyword{internal}
