% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep_diabetes_diagnoses}
\alias{keep_diabetes_diagnoses}
\title{Keep rows with diabetes diagnoses.}
\usage{
keep_diabetes_diagnoses(lpr2, lpr3)
}
\arguments{
\item{lpr2}{The output from \code{\link[=prepare_lpr2]{prepare_lpr2()}}.}

\item{lpr3}{The output from \code{\link[=prepare_lpr3]{prepare_lpr3()}}.}
}
\value{
The same type as the input data, as a \code{\link[duckplyr:duckdb_tibble]{duckplyr::duckdb_tibble()}},
with less rows after filtering.
}
\description{
This function uses the hospital contacts from LPR2 and LPR3 to include all
dates of diabetes diagnoses to use for inclusion, as well as
additional information needed to classify diabetes type.
Diabetes diagnoses from both ICD-8 and ICD-10 are included.
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
