% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep_pregnancy_dates}
\alias{keep_pregnancy_dates}
\title{Simple function to get only the pregnancy event dates.}
\usage{
keep_pregnancy_dates(lpr2, lpr3)
}
\arguments{
\item{lpr2}{Output from \code{\link[=prepare_lpr2]{prepare_lpr2()}}.}

\item{lpr3}{Output from \code{\link[=prepare_lpr3]{prepare_lpr3()}}.}
}
\value{
The same type as the input data, as a \code{\link[duckplyr:duckdb_tibble]{duckplyr::duckdb_tibble()}}.
}
\description{
Simple function to get only the pregnancy event dates.
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
