% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{logic_as_expr}
\alias{logic_as_expr}
\title{Get the algorithmic logic and convert to an R logic condition.}
\usage{
logic_as_expr(logic)
}
\arguments{
\item{logic_name}{The name of the logic to use.}

\item{algorithm}{The list of algorithmic logic, one list for each.}
}
\value{
An R expression.
}
\description{
Get the algorithmic logic and convert to an R logic condition.
}
\keyword{internal}
