% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{logic_as_expression}
\alias{logic_as_expression}
\title{Parse the logic strings into R expressions}
\usage{
logic_as_expression(logic)
}
\arguments{
\item{logic}{The name of the logic to use.}
}
\value{
An R expression.
}
\description{
Parse the logic strings into R expressions
}
\keyword{internal}
