% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-cases.R
\name{non_cases}
\alias{non_cases}
\title{List of non-cases to test the diabetes classification algorithm}
\usage{
non_cases()
}
\value{
A named list of 9 \code{\link[tibble:tibble]{tibble::tibble()}} objects, each representing a
different health register: \code{bef}, \code{lmdb}, \code{lpr_adm}, \code{lpr_diag},
\code{kontakter}, \code{diagnoser}, \code{sysi}, \code{sssy}, and \code{lab_forsker}.
}
\description{
This function generates a list of tibbles representing the Danish health
registers and the data necessary to run the algorithm. The dataset contains
individuals who should \emph{not} be included in the final classified cohort.
}
\details{
The generated data is used in \code{testthat} tests to ensure the algorithm
behaves as expected under a wide range of conditions, but it is also intended
to be explored by users to better understand how the algorithm logic works
and to be shown in the documentation.
}
\examples{
non_cases()
}
