% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osdc-package.R
\docType{package}
\name{osdc-package}
\alias{osdc}
\alias{osdc-package}
\title{osdc: Open Source Diabetes Classifier for Danish Registers}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The algorithm first identifies a population of individuals from Danish register data with any type of diabetes as individuals with two or more inclusion events. Then, it splits this population into individuals with either type 1 diabetes or type 2 diabetes by identifying individuals with type 1 diabetes and classifying the remainder of the diabetes population as having type 2 diabetes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/steno-aarhus/osdc}
  \item \url{https://steno-aarhus.github.io/osdc/}
  \item Report bugs at \url{https://github.com/steno-aarhus/osdc/issues}
}

}
\author{
\strong{Maintainer}: Luke William Johnston \email{lwjohnst@gmail.com} (\href{https://orcid.org/0000-0003-4169-2616}{ORCID})

Authors:
\itemize{
  \item Signe Kirk Brødbæk \email{signekb@clin.au.dk} (\href{https://orcid.org/0009-0000-2208-7088}{ORCID})
  \item Anders Aasted Isaksen \email{andaas@rm.dk} (\href{https://orcid.org/0000-0001-8457-5466}{ORCID})
}

Other contributors:
\itemize{
  \item Steno Diabetes Center Aarhus [copyright holder]
  \item Aarhus University [copyright holder]
}

}
\keyword{internal}
