% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{pad_integers}
\alias{pad_integers}
\title{Zero pad an integer to a specific length}
\usage{
pad_integers(x, width)
}
\arguments{
\item{x}{An integer or vector of integers.}

\item{width}{An integer describing the final width of the zero-padded integer.}
}
\value{
A character vector of integers.
}
\description{
Zero pad an integer to a specific length
}
\keyword{internal}
