% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-lpr.R
\name{prepare_lpr2}
\alias{prepare_lpr2}
\title{Prepare the two LPR2 registers into one to extract diabetes diagnoses data.}
\usage{
prepare_lpr2(lpr_adm, lpr_diag)
}
\arguments{
\item{lpr_adm}{The LPR2 register containing hospital admissions.}

\item{lpr_diag}{The LPR2 register containing diabetes diagnoses.}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}},
with the following columns:
\itemize{
\item \code{pnr}: The personal identification variable.
\item \code{date}: The date of all the recorded diagnosis (renamed from \code{d_inddto}).
\item \code{is_primary_diagnosis}: Whether the diagnosis was a primary diagnosis.
\item \code{has_diabetes}: Whether the diagnosis was any type of diabetes.
\item \code{has_t1d}: Whether the diagnosis was T1D-specific.
\item \code{has_t2d}: Whether the diagnosis was T2D-specific.
\item \code{has_pregnancy_event}: Whether the person has an event related to pregnancy like giving birth or having a miscarriage at the given date.
\item \code{is_endocrinology_department}: Whether the diagnosis was made made by an
endocrinology (TRUE) or other medical (FALSE) department.
}
}
\description{
The output is used as inputs to \code{include_diabetes_diagnoses()} and to
\code{get_pregnancy_dates()} (see exclusion events).
}
\examples{
\dontrun{
register_data <- simulate_registers(c("lpr_diag", "lpr_adm"), 100000)
lpr2 <- prepare_lpr2(
  lpr_diag = register_data$lpr_diag,
  lpr_adm = register_data$lpr_adm
)
}
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
