% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-lpr.R
\name{prepare_lpr_f}
\alias{prepare_lpr_f}
\title{Prepare and join the two LPR_F registers to extract diabetes and pregnancy diagnoses.}
\usage{
prepare_lpr_f(kontakter, diagnoser)
}
\arguments{
\item{kontakter}{The LPR_F register containing hospital contacts/administrative data.}

\item{diagnoser}{The LPR_F register containing diabetes diagnoses.}
}
\value{
The same type as the input data, as a \code{\link[duckplyr:duckdb_tibble]{duckplyr::duckdb_tibble()}},
with the following columns:
\itemize{
\item \code{pnr}: The personal identification variable.
\item \code{date}: The date of all the recorded diagnosis (renamed from
\code{d_inddto} or \code{dato_start}).
\item \code{is_primary_diagnosis}: Whether the diagnosis was a primary diagnosis.
\item \code{is_diabetes_code}: Whether the diagnosis was any type of diabetes.
\item \code{is_t1d_code}: Whether the diagnosis was T1D-specific.
\item \code{is_t2d_code}: Whether the diagnosis was T2D-specific.
\item \code{is_pregnancy_code}: Whether the person has an event related to
pregnancy like giving birth or having a miscarriage at the given date.
\item \code{is_endocrinology_dept}: Whether the diagnosis was made by an
endocrinology medical department.
\item \code{is_medical_dept}: Whether the diagnosis was made by a
non-endocrinology medical department.
}
}
\description{
Note that LPR_F is deprecated by the Health Data Authority, and should only
be used in projects where LPR_A does not contain all the LPR 3 data.
So this function should only be used in these cases, and care should be taken
to avoid duplicating rows where the coverage of LPR_A and LPR_F overlap.
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
