% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-markdown.R
\name{register_as_md_header}
\alias{register_as_md_header}
\title{Convert the register name into text to use in a Markdown header}
\usage{
register_as_md_header(register)
}
\arguments{
\item{register}{The abbreviation of the register name.}
}
\value{
A character vector.
}
\description{
Convert the register name into text to use in a Markdown header
}
\keyword{internal}
