% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osdc-package.R
\docType{data}
\name{register_data}
\alias{register_data}
\title{Simulation of register data needed for OSDC algorithm}
\format{
Is a list with several registers.
}
\usage{
register_data
}
\description{
Details of how this data was simulated is found in the \verb{data-raw/} folder on
the GitHub repository: \url{https://github.com/steno-aarhus/osdc}
}
\keyword{datasets}
