% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-markdown.R
\name{register_data_as_md_table}
\alias{register_data_as_md_table}
\title{Convert the fake register data into a Markdown table}
\usage{
register_data_as_md_table(data, caption = NULL)
}
\arguments{
\item{data}{The data of a specific register from \link{register_data}.}

\item{caption}{A caption to add to the table.}
}
\value{
A character vector as a Markdown table.
}
\description{
Convert the fake register data into a Markdown table
}
\keyword{internal}
