% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-markdown.R
\name{registers_as_md_table}
\alias{registers_as_md_table}
\title{Convert the register data sources}
\usage{
registers_as_md_table(caption = NULL)
}
\arguments{
\item{caption}{Caption to add to the table.}
}
\value{
A character vector as a Markdown table.
}
\description{
Convert the register data sources
}
\keyword{internal}
