% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{scrape_icd8_codes}
\alias{scrape_icd8_codes}
\title{Scrape SDS's website for their ICD-8 codes and saves to a file.}
\usage{
scrape_icd8_codes()
}
\value{
Saves a CSV file with ICD-8 codes. Outputs the path to the saved
file.
}
\description{
Scrape SDS's website for their ICD-8 codes and saves to a file.
}
\examples{
\dontrun{
scrape_icd8_codes()
}
}
\keyword{internal}
