% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_required_variables}
\alias{select_required_variables}
\title{Select the required variables from the register}
\usage{
select_required_variables(data, register)
}
\arguments{
\item{data}{The register to select columns from.}

\item{register}{The abbreviation of the register name. See list of
abbreviations in \code{\link[=get_register_abbrev]{get_register_abbrev()}}.}
}
\value{
Outputs the register with only the required variables, and
with column names in lower case.
}
\description{
This function selects only the required variables, convert to lower case,
and then check that the data types are as expected.
}
\keyword{internal}
