% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_registers}
\alias{simulate_registers}
\title{Simulate a fake data frame of one or more Danish registers}
\usage{
simulate_registers(registers, n = 1000)
}
\arguments{
\item{registers}{The name of the register you want to simulate.}

\item{n}{The number of rows to simulate for the resulting register.}
}
\value{
A list with simulated register data.
}
\description{
Simulate a fake data frame of one or more Danish registers
}
\examples{
simulate_registers(c("bef", "sysi"))
simulate_registers("bef")
}
