% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{to_yyww}
\alias{to_yyww}
\title{Transform date(s) to the format \code{yyww}}
\usage{
to_yyww(x)
}
\arguments{
\item{x}{A date or a vector of dates.}
}
\value{
A vector of dates in the format \code{yyww}.
}
\description{
Transform date(s) to the format \code{yyww}
}
\examples{
\dontrun{
to_yyww("2020-12-01")
to_yyww("2001-01-01")
to_yyww(c("2020-01-12", "1995-04-19"))
}
}
\keyword{internal}
