% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{to_yyyymmdd}
\alias{to_yyyymmdd}
\title{Transform date(s) to the format \code{yyyymmdd}}
\usage{
to_yyyymmdd(x)
}
\arguments{
\item{x}{A date or a vector of dates.}
}
\value{
A vector of dates in the format \code{yyyymmdd.}
}
\description{
Transform date(s) to the format \code{yyyymmdd}
}
\examples{
\dontrun{
to_yyyymmdd("2020-12-01")
to_yyyymmdd(c("2020-01-12", "1995-04-19"))
}
}
\keyword{internal}
