% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-description.R
\docType{data}
\name{variable_description}
\alias{variable_description}
\title{Variables from registers and their descriptions that are required for the
osdc algorithm.}
\format{
\subsection{\code{variable_description}}{

A data frame with 39 rows and 6 columns:
\describe{
\item{register_name}{The official, full Danish name of the register.}
\item{register_abbrev}{The official abbreviation for the register.}
\item{variable_name}{The official name of the variable found in the register.}
\item{years_covered}{The years when the variable is available from.}
\item{danish_description}{The official description in Danish for the variable.}
\item{english_description}{The translated description in English for the variable.}
}
}
}
\source{
Many of the details within the \code{variable_description} dataset come
from the full official list of registers from Statistics Denmark (DST):
\url{https://www.dst.dk/extranet/forskningvariabellister/Oversigt\%20over\%20registre.html}
}
\usage{
variable_description
}
\description{
Variables from registers and their descriptions that are required for the
osdc algorithm.
}
\keyword{datasets}
