% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-markdown.R
\name{variables_as_md_table}
\alias{variables_as_md_table}
\title{Converts the variables for a register into a Markdown table}
\usage{
variables_as_md_table(register, caption = NULL)
}
\arguments{
\item{caption}{A caption to add to the table.}
}
\value{
A character vector as a Markdown table.
}
\description{
Converts the variables for a register into a Markdown table
}
\keyword{internal}
