% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyww-to-date.R
\name{yyww_to_date}
\alias{yyww_to_date}
\title{Convert date format YYWW to YYYY-MM-DD}
\usage{
yyww_to_date(yyww)
}
\arguments{
\item{yyww}{Character(s) of the format yyww}
}
\value{
Date(s) in the format YYYY-MM-DD.
}
\description{
Since the exact date isn't given in the input, this function will set the
date to Monday of the week.
}
\details{
As a precaution, a leading zero is added if it has been removed
This can e.g., happen if the input was "0107" and has been converted to a
numeric 107.
}
\examples{
\dontrun{
yyww_to_date("0102")
yyww_to_date(c("0102", "0304"))
}
}
\keyword{internal}
