% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{yyww_to_yyyymmdd}
\alias{yyww_to_yyyymmdd}
\title{Convert date format YYWW to YYYY-MM-DD}
\usage{
yyww_to_yyyymmdd(data)
}
\arguments{
\item{data}{A DuckDB-backed data frame with a \code{date} column formatted as YYWW values.}
}
\value{
a data frame containing a \code{date} column in the format YYYY-MM-DD.
}
\description{
Since the exact date isn't given in the input, this function will set the
date to Monday of the week. As a precaution, a leading zero is added if it
has been removed. This can e.g., happen if the input was "0107" and has been
converted to a numeric 107.
}
\keyword{internal}
