% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_insulin_purchases_cols}
\alias{add_insulin_purchases_cols}
\title{Add three columns related to insulin purchases}
\usage{
add_insulin_purchases_cols(data, gld_purchases)
}
\arguments{
\item{data}{A \code{\link[tibble:tibble]{tibble::tibble()}} with at least the column \code{pnr}.}

\item{gld_purchases}{The output from \code{\link[=include_gld_purchases]{include_gld_purchases()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with three new logical columns called
\code{is_two_thirds_insulin}, \code{is_only_insulin_purchases}, and
\code{is_insulin_purchases_within_180_days}.
}
\description{
The columns are 1) if an insulin purchase has been made within 180 days of
the first glucose-lowering drug purchase, 2) if two-thirds of purchased
glucose-lowering drug doses are insulin, and 3) if insulin is the only
glucose-lowering drug type ever purchased.
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
