% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osdc-package.R
\docType{data}
\name{algorithm}
\alias{algorithm}
\title{Data frame of the logic for the OSDC algorithm}
\format{
Is a \code{\link[tibble:tibble]{tibble::tibble()}} with two columns:

\describe{
\item{name}{The inclusion or exclusion criteria name.}
\item{logic}{The logic for the criteria.}
}
}
\usage{
algorithm
}
\description{
This data frame contains the logic details of the algorithm for specific
inclusion and exclusion criteria.
}
\keyword{datasets}
