% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-required-variables.R
\name{check_data_types}
\alias{check_data_types}
\title{Check data types of the register variables}
\usage{
check_data_types(data, register, call = rlang::caller_env())
}
\value{
Either TRUE if the verification passes, or an error.
}
\description{
Check data types of the register variables
}
\examples{
\dontrun{
check_data_types(simulate_registers("kontakter")[[1]], "kontakter")
}
}
\keyword{internal}
