% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{check_data_types}
\alias{check_data_types}
\title{Check data types of the register variables}
\usage{
check_data_types(data, register, call = rlang::caller_env())
}
\value{
Outputs the register with only the required variables, and
with column names in lower case.
}
\description{
Check data types of the register variables
}
\examples{
\dontrun{
check_data_types(simulate_registers("kontakter")[[1]], "kontakter")
}
}
\keyword{internal}
