% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-diabetes.R
\name{classify_diabetes}
\alias{classify_diabetes}
\title{Classify diabetes status using Danish registers.}
\usage{
classify_diabetes(
  kontakter,
  diagnoser,
  lpr_diag,
  lpr_adm,
  sysi,
  sssy,
  lab_forsker,
  bef,
  lmdb,
  stable_inclusion_start_date = "1998-01-01"
)
}
\arguments{
\item{kontakter}{The contacts information table from the lpr3 patient register}

\item{diagnoser}{The diagnoses information table from the lpr3 patient register}

\item{lpr_diag}{The diagnoses information table from the lpr2 patient register}

\item{lpr_adm}{The administrative information table from the lpr2 patient register}

\item{sysi}{The SYSI table from the health service register}

\item{sssy}{The SSSY table from the health service register}

\item{lab_forsker}{The register for laboratory results for research}

\item{bef}{The BEF table from the civil register}

\item{lmdb}{The LMDB table from the prescription register}

\item{stable_inclusion_start_date}{Cutoff date after which inclusion events
are considered true incident diabetes cases. Defaults to "1998-01-01", as
we presume the user is using data on pregnancy events from the the Patient Register
and on purchases of glucose-lowering drugs from the Prescription Register
from Jan 1 1997 onward (or earlier).}
}
\value{
The same object type as the input data, which would be a
\code{\link[duckplyr:duckdb_tibble]{duckplyr::duckdb_tibble()}} type object.
}
\description{
This function requires that each source of raw data is represented
as a single DuckDB object in R (e.g. a connection to Parquet files).
Each DuckDB object must contain a single table covering all years of
that data source, or at least the years you have and are interested
in.
}
\examples{
\dontrun{
register_data <- simulate_registers(
  c(
    "kontakter",
    "diagnoser",
    "lpr_diag",
    "lpr_adm",
    "sysi",
    "sssy",
    "lab_forsker",
    "bef",
    "lmdb"
  ),
  n = 10000
)
classify_diabetes(
  kontakter = register_data$kontakter,
  diagnoser = register_data$diagnoser,
  lpr_diag = register_data$lpr_diag,
  lpr_adm = register_data$lpr_adm,
  sysi = register_data$sysi,
  sssy = register_data$sssy,
  lab_forsker = register_data$lab_forsker,
  bef = register_data$bef,
  lmdb = register_data$lmdb
)
}
}
\seealso{
See the \link{osdc} vignette for a detailed
description of the internal implementation of this classification function.
}
