% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-diabetes.R
\name{classify_diabetes}
\alias{classify_diabetes}
\title{Classify diabetes status using Danish registers.}
\usage{
classify_diabetes(
  kontakter,
  diagnoser,
  lpr_diag,
  lpr_adm,
  sysi,
  sssy,
  lab_forsker,
  bef,
  lmdb,
  stable_inclusion_start_date = "1998-01-01"
)
}
\arguments{
\item{kontakter}{The contacts information register for lpr3}

\item{diagnoser}{The diagnoses information register for lpr3}

\item{lpr_diag}{The diagnoses information register for lpr2}

\item{lpr_adm}{The administrative information register for lpr2}

\item{sysi}{The SYSI register}

\item{sssy}{The SSSY register}

\item{lab_forsker}{The laboratory results for research register}

\item{bef}{The BEF register}

\item{lmdb}{The LMDB register}

\item{stable_inclusion_start_date}{Cutoff date after which inclusion events
are considered reliable (e.g., after changes in drug labeling or data
entry practices). Defaults to "1998-01-01" which is one year after
obstetric codes are reliable in the GLD data (since we use LPR data to
drop rows related to gestational diabetes). This limits the included
cohort to individuals with inclusion dates after this cutoff date.}
}
\value{
The same object type as the input data, which would be a
\code{\link[tibble:tibble]{tibble::tibble()}} type object.
}
\description{
This function assumes that each source of raw data is represented as a single
object in R, e.g. a DuckDB connection to the parquet file(s) has been set up
or the data has already been read into memory.
}
\details{
This section should probably describe the expected structure of each data
source (e.g. a list of tables or a single table)
}
\examples{
\dontrun{
register_data <- simulate_registers(
  c(
    "kontakter",
    "diagnoser",
    "lpr_diag",
    "lpr_adm",
    "sysi",
    "sssy",
    "lab_forsker",
    "bef",
    "lmdb"
  ),
  n = 10000
)
classify_diabetes(
  kontakter = register_data$kontakter,
  diagnoser = register_data$diagnoser,
  lpr_diag = register_data$lpr_diag,
  lpr_adm = register_data$lpr_adm,
  sysi = register_data$sysi,
  sssy = register_data$sssy,
  lab_forsker = register_data$lab_forsker,
  bef = register_data$bef,
  lmdb = register_data$lmdb
)
}
}
\seealso{
See the \link{osdc} vignette for a detailed
description of the internal implementation of this classification function.
}
