% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-diabetes.R
\name{classify_diabetes}
\alias{classify_diabetes}
\title{Classify diabetes status using Danish registers.}
\usage{
classify_diabetes(
  kontakter,
  diagnoser,
  lpr_diag,
  lpr_adm,
  sysi,
  sssy,
  lab_forsker,
  bef,
  lmdb
)
}
\value{
The same object type as the input data, which would be a
\code{\link[tibble:tibble]{tibble::tibble()}} type object.
}
\description{
Classify diabetes status using Danish registers.
}
\examples{
\dontrun{
classify_diabetes(
  kontakter = register_data$kontakter,
  diagnoser = register_data$diagnoser,
  lpr_diag = register_data$lpr_diag,
  lpr_adm = register_data$lpr_adm,
  sysi = register_data$sysi,
  sssy = register_data$sssy,
  lab_forsker = register_data$lab_forsker,
  bef = register_data$bef,
  lmdb = register_data$lmdb
)
}
}
\seealso{
See the \link{vignette("function-flow", package = "osdc")} for a detailed
description of the internal implementation of this classification function.
}
