% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{column_names_to_lower}
\alias{column_names_to_lower}
\title{Convert column names to lower case}
\usage{
column_names_to_lower(data)
}
\arguments{
\item{data}{An data frame type object.}
}
\value{
The same object type given.
}
\description{
Convert column names to lower case
}
\keyword{internal}
