% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{create_padded_integer}
\alias{create_padded_integer}
\title{Create a vector of reproducible, random zero-padded integers}
\usage{
create_padded_integer(n, length)
}
\arguments{
\item{n}{The number of integers to generate.}

\item{length}{An integer determining the length of the padded integer.}
}
\value{
A character vector of integers.
}
\description{
Generated integers of the same length are identical to facilitate joining by
values in \code{pnr}, \code{cpr}, \code{recnum} and \code{dw_ek_kontakt}.
}
\examples{
\dontrun{
create_padded_integer(5, 10)
}
}
\keyword{internal}
