% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{get_algorithm_logic}
\alias{get_algorithm_logic}
\title{Get the algorithmic logic and convert to an R logic condition.}
\usage{
get_algorithm_logic(logic_name, algorithm = NULL)
}
\arguments{
\item{logic_name}{The name of the logic to use.}

\item{algorithm}{The list of algorithmic logic, one list for each.}
}
\value{
A character string.
}
\description{
Get the algorithmic logic and convert to an R logic condition.
}
\examples{
\dontrun{
get_algorithm_logic("hba1c")
get_algorithm_logic("gld")
}
}
\keyword{internal}
