% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-diabetes-diagnoses.R
\name{include_diabetes_diagnoses}
\alias{include_diabetes_diagnoses}
\title{Determine if majority of diagnoses are type 1 diabetes}
\usage{
include_diabetes_diagnoses(lpr2, lpr3)
}
\arguments{
\item{lpr2}{The output from \code{\link[=prepare_lpr2]{prepare_lpr2()}}.}

\item{lpr3}{The output from \code{\link[=prepare_lpr3]{prepare_lpr3()}}.}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}},
with less rows after filtering.
}
\description{
Uses the hospital contacts from LPR2 and LPR3 to include all
dates of diabetes diagnoses to use for inclusion, as well as
additional information needed to classify diabetes type.
Diabetes diagnoses from both ICD-8 and ICD-10 are included.
}
\examples{
\dontrun{
register_data <- simulate_registers(c("lpr_diag", "lpr_adm", "diagnoser", "kontakter"))
include_diabetes_diagnoses(
  lpr2 = prepare_lpr2(register_data$lpr_adm, register_data$lpr_diag),
  lpr3 = prepare_lpr3(register_data$kontakter, register_data$diagnoser)
)
}
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
