% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-diabetes-diagnoses.R
\name{include_diabetes_diagnoses}
\alias{include_diabetes_diagnoses}
\title{Include ICD-8 and ICD-10 codes of diabetes from hospital contacts for inclusion and type-specification.}
\usage{
include_diabetes_diagnoses(lpr2_diag, lpr2_adm, lpr3_diagnoser, lpr3_kontakter)
}
\arguments{
\item{lpr_diag}{The diagnosis table from LPR2}

\item{lpr_adm}{The admission table from LPR2}

\item{diagnoser}{The diagnosis table from LPR3}

\item{kontakter}{The contacts table from LPR3}
}
\value{
An object of the same input type, default as a \code{\link[tibble:tibble]{tibble::tibble()}},
with six columns and up to two rows per unique \code{pnr} value):
\itemize{
\item \code{pnr}: identifier variable
\item \code{dates}: dates of the first and second hospital diabetes diagnosis
\item \code{n_t1d_endocrinology}: number of type 1 diabetes-specific primary diagnosis codes from endocrinological departments
\item \code{n_t2d_endocrinology}: number of type 2 diabetes-specific primary diagnosis codes from endocrinological departments
\item \code{n_t1d_medical}: number of type 1 diabetes-specific primary diagnosis codes from medical departments
\item \code{n_t2d_medical}: number of type 2 diabetes-specific primary diagnosis codes from medical departments
This output is passed to the \code{join_inclusions()} function, where the \code{dates} variable is used for the final step of the inclusion process.
The variables of counts of diabetes type-specific primary diagnoses (the four columns prefixed \code{n_} above) are carried over for the subsequent classification of diabetes type, initially as inputs to the \code{get_t1d_primary_diagnosis()} and \code{get_majority_of_t1d_diagnoses()} functions..
}
}
\description{
This function calls \code{join_lpr2()} and \code{join_lpr3()} and processes their output for inclusion, as well as additional information needed to classify diabetes type.
}
\examples{
\dontrun{
include_diabetes_diagnoses(lpr_diag, lpr_adm, diagnoser, kontakter)
}
}
\keyword{internal}
