% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-diabetes-diagnoses.R
\name{include_diabetes_diagnoses}
\alias{include_diabetes_diagnoses}
\title{Include ICD-8 and ICD-10 Codes for Diabetes Diagnosis and Type-Specific Classification}
\usage{
include_diabetes_diagnoses(lpr_diag, lpr_adm, diagnoser, kontakter)
}
\arguments{
\item{lpr_diag}{The diagnosis table from LPR2}

\item{lpr_adm}{The admission table from LPR2}

\item{diagnoser}{The diagnosis table from LPR3}

\item{kontakter}{The contacts table from LPR3}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{pnr}: Identifier variable (unique to individuals).
\item \code{date}: Dates of the first and second hospital diabetes diagnoses.
\item \code{any_t1d_primary_diagnosis}: Logical value indicating whether an individual has at least one type 1 diabetes-specific primary diagnosis code from either an endocrinological or other medical department.
\item \code{majority_t1d_primary_diagnoses}: Logical value indicating whether an individual has a majority of type 1 diabetes-specific primary diagnoses among all type-specific primary diagnoses recorded in endocrinological departments (or among diagnoses from other medical departments if no endocrinological contacts are recorded).
}

The output is passed to \code{join_inclusions()}, where the \code{date} variable is used to define the inclusion date.
The variables \code{any_t1d_primary_diagnosis} and \code{majority_t1d_primary_diagnoses} are subsequently passed to \code{get_diabetes_type()} for final classification of diabetes type.
}
\description{
This function processes hospital contact data by combining outputs from \code{join_lpr2()} and \code{join_lpr3()} to identify diabetes-related diagnoses, evaluate inclusion criteria, and compute variables for type-specific classification.
}
\examples{
\dontrun{
include_diabetes_diagnoses(lpr_diag, lpr_adm, diagnoser, kontakter)
}
}
\keyword{internal}
