% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-gld-purchases.R
\name{include_gld_purchases}
\alias{include_gld_purchases}
\title{Include only those who have a purchase of a glucose lowering drug (GLD).}
\usage{
include_gld_purchases(lmdb)
}
\arguments{
\item{lmdb}{The \code{lmdb} register.}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}}, in
a long format with all dates of GLD purchases kept and the following
variables:
\itemize{
\item \code{pnr}: Personal identification variable.
\item \code{date}: The dates of all purchases of GLD.
\item \code{atc}: The ATC code for the type of drug.
\item \code{indication_code}: The indication code of the prescription (renamed
from \code{indo}).
\item \code{has_two_thirds_insulin}: A logical variable used in classifying type 1
diabetes. See \code{\link[=algorithm]{algorithm()}} for more detail.
\item \code{has_only_insulin_purchases}: A logical variable used in classifying type 1
diabetes. See \code{\link[=algorithm]{algorithm()}} for more detail.
\item \code{has_insulin_purchases_within_180_days}: A logical variable used in classifying type 1
diabetes. See \code{\link[=algorithm]{algorithm()}} for more detail.
}
}
\description{
But don't include glucose-lowering drugs that may be used for other
conditions than diabetes like GLP-RAs or dapagliflozin/empagliflozin drugs.
Since the diagnosis code data on pregnancies (see below) is insufficient to
perform censoring prior to 1997, \code{include_gld_purchases()} only extracts
dates from 1997 onward by default (if Medical Birth Register data is
available to use for censoring, the extraction window can be extended).
}
\details{
Additional columns added are 1) if an insulin purchase has been made within
180 days of the first glucose-lowering drug purchase, 2) if two-thirds of
purchased glucose-lowering drug doses are insulin, and 3) if insulin is the
only glucose-lowering drug type ever purchased.

These events are then passed to a chain of exclusion functions:
\code{exclude_potential_pcos()} and \code{exclude_pregnancy()}.
}
\examples{
\dontrun{
simulate_registers("lmdb", 10000)[[1]] |> include_gld_purchases()
}
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
