% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-hba1c.R
\name{include_hba1c}
\alias{include_hba1c}
\title{Include only those with HbA1c in the required range.}
\usage{
include_hba1c(data)
}
\arguments{
\item{data}{The \code{lab_forsker} register.}
}
\value{
An object of the same input type, default as a \code{\link[tibble:tibble]{tibble::tibble()}},
with two columns: \code{pnr} and \code{included_hba1c}.
}
\description{
In the \code{lab_forsker} register, NPU27300 is HbA1c in the modern units (IFCC)
while NPU03835 is HbA1c in old units (DCCT).
}
\examples{
\dontrun{
register_data$lab_forsker |> include_hba1c()
}
}
\keyword{internal}
