% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-podiatrist-services.R
\name{include_podiatrist_services}
\alias{include_podiatrist_services}
\title{Include diabetes-specific podiatrist services.}
\usage{
include_podiatrist_services(sysi, sssy)
}
\arguments{
\item{sysi}{The SYSI register.}

\item{sssy}{The SSSY register.}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}},
with two columns and up to two rows for each individual:
\itemize{
\item \code{pnr}: Identifier variable
\item \code{date}: The dates of the first and second diabetes-specific
podiatrist record
\item \code{has_podiatrist_services}: A logical variable that acts as a helper
indicator for use in later functions.
}
}
\description{
#' Uses the \code{sysi} or \code{sssy} registers as input to extract the dates of all
diabetes-specific podiatrist services. Removes duplicate services on the
same date. Only the two earliest dates per individual are kept.
}
\details{
#' The output is passed to the \code{join_inclusions()} function for the final
step of the inclusion process.
}
\examples{
\dontrun{
include_podiatrist_services(register_data$sysi, register_data$sssy)
}
}
\keyword{internal}
