% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{join_lpr2}
\alias{join_lpr2}
\title{Join together the LPR2 (\code{lpr_diag} and \code{lpr_adm}) registers.}
\usage{
join_lpr2(lpr_diag, lpr_adm)
}
\arguments{
\item{lpr_diag}{The diagnosis register.}

\item{lpr_adm}{The admission register.}
}
\value{
The same class as the input, defaults to a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Join together the LPR2 (\code{lpr_diag} and \code{lpr_adm}) registers.
}
\examples{
register_data$lpr_diag |>
  join_lpr2(register_data$lpr_adm)
}
\keyword{internal}
