% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep_pregnancy_dates}
\alias{keep_pregnancy_dates}
\title{Simple function to get only the pregnancy event dates.}
\usage{
keep_pregnancy_dates(lpr2, lpr3)
}
\arguments{
\item{lpr2}{Output from \code{prepare_lpr2()}.}

\item{lpr3}{Output from \code{prepare_lpr3()}.}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Simple function to get only the pregnancy event dates.
}
\examples{
\dontrun{
register_data <- simulate_registers(
  c("lpr_adm", "lpr_diag", "kontakter", "diagnoser"),
  n = 1000
)
lpr2 <- prepare_lpr2(register_data$lpr_adm, register_data$lpr_diag)
lpr3 <- prepare_lpr3(register_data$diagnoser, register_data$kontakter)
keep_pregnancy_dates(lpr2, lpr3)
}
}
\seealso{
See the \code{vignette("algorithm")} for the logic used to filter these
patients.
}
\keyword{internal}
