% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep_two_earliest_events}
\alias{keep_two_earliest_events}
\title{Keep two earliest events per PNR}
\usage{
keep_two_earliest_events(data)
}
\arguments{
\item{data}{Data including at least a date and pnr column.}
}
\value{
The same type as the input data, default as a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Since the classification date is based on the second instance of
an inclusion criteria, we only need to keep the earliest two events per PNR
per inclusion "stream".
}
\details{
This function is applied to each "stream", \code{diabetes_diagnoses},
\code{podiatrist_services}, and \code{gld_hba1c_after_drop_steps}, in  the
\code{classify_diabetes()} function after the keep and drop steps, right before
they are joined.
}
\examples{
\dontrun{
data <- tibble::tribble(
  ~pnr, ~date,
  1, "20200101",
  1, "20200102",
  1, "20200103",
  2, "20200101"
)
keep_two_earliest_events(data)
}
}
\keyword{internal}
