% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-cases.R
\name{non_cases_metadata}
\alias{non_cases_metadata}
\title{Description of the different non-cases included in \code{non_cases()}}
\usage{
non_cases_metadata()
}
\value{
A named list of character strings, where each name corresponds to a
non-case PNR in the dataset generated by \code{non_cases()}.
}
\description{
All cases, aside from what would exclude them from being classified as
described in the metadata here, would otherwise be classified as having
diabetes.
}
\examples{
non_cases_metadata()
}
