% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osdc-package.R
\docType{package}
\name{osdc-package}
\alias{osdc}
\alias{osdc-package}
\title{osdc: Open Source Diabetes Classifier (OSCD) for Danish Registers}
\description{
This classifier first identifies a population of individuals with any type of diabetes mellitus and then splits this population into T1D and T2D by identifying individuals with T1D and classifying the remainder of the diabetes population as T2D.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/steno-aarhus/osdc}
  \item \url{https://steno-aarhus.github.io/osdc/}
  \item Report bugs at \url{https://github.com/steno-aarhus/osdc/issues}
}

}
\author{
\strong{Maintainer}: Luke William Johnston \email{lwjohnst@gmail.com} (\href{https://orcid.org/0000-0003-4169-2616}{ORCID})

Authors:
\itemize{
  \item Signe Kirk Brødbæk \email{signekb@clin.au.dk}
  \item Anders Aasted Isaksen \email{andaas@rm.dk}
}

Other contributors:
\itemize{
  \item Steno Diabetes Center Aarhus [copyright holder]
  \item Aarhus University [copyright holder]
}

}
\keyword{internal}
