% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_required_variables}
\alias{select_required_variables}
\title{Check that the dataset has the required variables for the algorithm.}
\usage{
select_required_variables(data, register, call = rlang::caller_env())
}
\arguments{
\item{data}{The dataset to check.}

\item{register}{The abbreviation of the register name. See list of
abbreviations in \code{\link[=get_register_abbrev]{get_register_abbrev()}}.}

\item{call}{The environment where the function is called, so that the error
traceback gives a more meaningful location.}
}
\value{
Either TRUE if the verification passes, or an error.
}
\description{
Use this function inside another function within an \code{if} condition to provide an
informative error message within the function used. This is done to make the
error message more informative to the location that the error actually
occurs, rather than within this function.
}
\examples{
\dontrun{
# TODO: Replace with simulated data.
select_required_variables(simulate_registers("bef")[[1]], "bef")
select_required_variables(simulate_registers("lpr_adm")[[1]], "lpr_adm")
}
}
\keyword{internal}
