% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_required_variables}
\alias{select_required_variables}
\title{Select the required variables from the register}
\usage{
select_required_variables(data, register, call = rlang::caller_env())
}
\arguments{
\item{data}{The dataset to check.}

\item{register}{The abbreviation of the register name. See list of
abbreviations in \code{\link[=get_register_abbrev]{get_register_abbrev()}}.}

\item{call}{The environment where the function is called, so that the error
traceback gives a more meaningful location.}
}
\value{
Outputs the register with only the required variables, and
with column names in lower case.
}
\description{
Use this function to select only the required variables, convert to lower case,
and then check that the data types are as expected.
}
\examples{
\dontrun{
select_required_variables(simulate_registers("bef")[[1]], "bef")
select_required_variables(simulate_registers("lpr_adm")[[1]], "lpr_adm")
}
}
\keyword{internal}
