% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{to_yyww}
\alias{to_yyww}
\title{Transform date(s) to the format \code{yyww}}
\usage{
to_yyww(x)
}
\arguments{
\item{x}{A date or a vector of dates.}
}
\value{
A vector of dates in the format \code{yyww}.
}
\description{
Transform date(s) to the format \code{yyww}
}
\keyword{internal}
