% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-registers.R
\name{to_yyyymmdd}
\alias{to_yyyymmdd}
\title{Transform date(s) to the format \code{yyyymmdd}}
\usage{
to_yyyymmdd(x)
}
\arguments{
\item{x}{A date or a vector of dates.}
}
\value{
A vector of dates in the format \code{yyyymmdd.}
}
\description{
Transform date(s) to the format \code{yyyymmdd}
}
\keyword{internal}
