% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify-variables.R
\name{verify_required_variables}
\alias{verify_required_variables}
\title{Verify that the dataset has the required variables for the algorithm.}
\usage{
verify_required_variables(data, register)
}
\arguments{
\item{data}{The dataset to check.}

\item{register}{The abbreviation of the register name. See list of
abbreviations in \code{\link[=get_register_abbrev]{get_register_abbrev()}}.}
}
\value{
Either TRUE if the verification passes, or a character string if
there is an error.
}
\description{
Use this function inside another function within an \code{if} condition to provide an
informative error message within the function used. This is done to make the
error message more informative to the location that the error actually
occurs, rather than within this function.
}
\examples{
# TODO: Replace with simulated data.
example_bef_data <- tibble::tibble(pnr = 1, koen = 1, foed_dato = 1)
osdc:::verify_required_variables(example_bef_data, "bef")
}
\keyword{internal}
